requirements.txt
requests==2.31.0
couchdb==1.2
python-dotenv==1.0.0
streamlit==1.28.1
matplotlib==3.8.2
plotly==5.17.0
pandas==2.1.4
pytest==7.4.3
jupyter==1.0.0
ipykernel==6.26.0